Removing serializable structures: image creation and restoring

+++ libdoclone/examples/doclone.cc
@@ -52,9 +52,9 @@ public:
 			Doclone::dcOperationType type, const std::string &target);
 	void notify(Doclone::dcEvent event, const std::string &target);
 	void notify(const std::string &str);
-	
+
 	void initView(int argc, char **argv) const;
-	
+
 private:
 	// Total amount of bytes to be transferred (for calculate the percentage)
 	uint64_t _totalSize;
@@ -82,7 +82,7 @@ void ConsoleView::notify(Doclone::dcTransferEvent event, const uint64_t numBytes
 	case Doclone::TRANS_TRANSFERRED_BYTES: {
 		if(this->_totalSize>0) {
 			// Calculate and print the percentage of data transferred
-			float rate = numBytes/static_cast<float>(this->_totalSize);
+			double rate = numBytes/static_cast<double>(this->_totalSize);
 			int percent = floor(rate*100);
 			if(this->_percent != percent) {
 				this->_percent = percent;
@@ -354,7 +354,7 @@ void ConsoleView::initView(int argc, char **argv) const {
 			break;
 		}
 		case 'f': {
-			if (!strrchr (optarg, '/'))	{ // If it is a relative path 
+			if (!strrchr (optarg, '/'))	{ // If it is a relative path
 				char tmp[256];
 				snprintf (tmp, sizeof(tmp), "./%s", optarg);
 				image = tmp;
@@ -482,7 +482,7 @@ void ConsoleView::initView(int argc, char **argv) const {
 
 int main(int argc, char **argv) {
 	ConsoleView view;
-	
+
 	view.initView(argc, argv);
 
 	return 0;
